HW2/requirements.txt
groq
pandas
python-dotenv
g4f[all]
nest_asyncio
click